+++ unpythonic/__init__.py
@@ -36,7 +36,6 @@ from .lispylet import (let as ordered_let, letrec as ordered_letrec,  # noqa: F4
 from .llist import *  # noqa: F401, F403
 from .mathseq import *  # noqa: F401, F403
 from .misc import *  # noqa: F401, F403
-from .numutil import *  # noqa: F401, F403
 from .seq import *  # noqa: F401, F403
 from .singleton import *  # noqa: F401, F403
 from .slicing import *  # noqa: F401, F403
@@ -58,3 +57,8 @@ from .funutil import _init_module
 _init_module()
 del _init_module
 from .funutil import *  # noqa: F401, F403
+
+from .numutil import _init_module
+_init_module()
+del _init_module
+from .numutil import *  # noqa: F401, F403