+++ unpythonic/it.py
@@ -23,10 +23,9 @@ __all__ = ["rev", "map", "map_longest",
            "flatten", "flatten1", "flatten_in",
            "iterate", "iterate1",
            "partition",
-           "partition_int",
            "inn", "iindex", "find",
            "window", "chunked",
-           "within", "fixpoint",
+           "within",
            "interleave",
            "subset", "powerset",
            "allsame"]
@@ -594,70 +593,13 @@ def partition(pred, iterable):
     It will eventually run out of memory storing all the odd numbers "to be read
     later".)
 
-    Not to be confused with `unpythonic.it.partition_int`, which partitions
+    Not to be confused with `unpythonic.numutil.partition_int`, which partitions
     a (small) positive integer to smaller integers, in all possible ways,
     such that those integers sum to the original one.
     """
     # iterable is walked only once; tee handles the intermediate storage.
     t1, t2 = tee(iterable)
     return filterfalse(pred, t1), filter(pred, t2)
-
-def partition_int(n, lower=1, upper=None):
-    """Yield all ordered sequences of smaller positive integers that sum to `n`.
-
-    `n` must be an integer >= 1.
-
-    `lower` is an optional lower limit for each member of the sum. Each member
-    of the sum must be `>= lower`.
-
-    (Most of the splits are a ravioli consisting mostly of ones, so it is much
-    faster to not generate such splits than to filter them out from the result.
-    The default value `lower=1` generates everything.)
-
-    `upper` is, similarly, an optional upper limit; each member of the sum
-    must be `<= upper`. The default `None` means no upper limit (effectively,
-    in that case `upper=n`).
-
-    It must hold that `1 <= lower <= upper <= n`.
-
-    Not to be confused with `unpythonic.it.partition`, which partitions an
-    iterable based on a predicate.
-
-    **CAUTION**: The number of possible partitions grows very quickly with `n`,
-    so in practice this is only useful for small numbers, or with a lower limit
-    that is not too much smaller than `n / 2`. A possible use case for this
-    function is to determine the number of letters to allocate for each
-    component of an anagram that may consist of several words.
-
-    See:
-        https://en.wikipedia.org/wiki/Partition_(number_theory)
-    """
-    # sanity check the preconditions, fail-fast
-    if not isinstance(n, int):
-        raise TypeError(f"n must be integer; got {type(n)} with value {repr(n)}")
-    if not isinstance(lower, int):
-        raise TypeError(f"lower must be integer; got {type(lower)} with value {repr(lower)}")
-    if upper is not None and not isinstance(upper, int):
-        raise TypeError(f"upper must be integer; got {type(upper)} with value {repr(upper)}")
-    upper = upper if upper is not None else n
-    if n < 1:
-        raise ValueError(f"n must be positive; got {n}")
-    if lower < 1 or upper < 1 or lower > n or upper > n or lower > upper:
-        raise ValueError(f"it must hold that 1 <= lower <= upper <= n; got lower={lower}, upper={upper}")
-
-    def _partition(n):
-        for k in range(min(n, upper), lower - 1, -1):
-            m = n - k
-            if m == 0:
-                yield (k,)
-            else:
-                out = []
-                for item in _partition(m):
-                    out.append((k,) + item)
-                for term in out:
-                    yield term
-
-    return _partition(n)  # instantiate the generator
 
 def inn(x, iterable):
     """Contains-check (``x in iterable``) with automatic termination.
@@ -839,42 +781,6 @@ def within(tol, iterable):
             yield b
             return
 
-def fixpoint(f, x0, tol=0):
-    """Compute the (arithmetic) fixed point of f, starting from the initial guess x0.
-
-    (Not to be confused with the logical fixed point with respect to the
-    definedness ordering.)
-
-    The fixed point must be attractive for this to work. See the Banach
-    fixed point theorem.
-    https://en.wikipedia.org/wiki/Banach_fixed-point_theorem
-
-    If the fixed point is attractive, and the values are represented in
-    floating point (hence finite precision), the computation should
-    eventually converge down to the last bit (barring roundoff or
-    catastrophic cancellation in the final few steps). Hence the default tol
-    of zero.
-
-    CAUTION: an arbitrary function from ℝ to ℝ **does not** necessarily
-    have a fixed point. Limit cycles and chaotic behavior of `f` will cause
-    non-termination. Keep in mind the classic example:
-    https://en.wikipedia.org/wiki/Logistic_map
-
-    Examples::
-        from math import cos, sqrt
-        from unpythonic import fixpoint, ulp
-        c = fixpoint(cos, x0=1)
-
-        # Actually "Newton's" algorithm for the square root was already known to the
-        # ancient Babylonians, ca. 2000 BCE. (Carl Boyer: History of mathematics)
-        def sqrt_newton(n):
-            def sqrt_iter(x):  # has an attractive fixed point at sqrt(n)
-                return (x + n / x) / 2
-            return fixpoint(sqrt_iter, x0=n / 2)
-        assert abs(sqrt_newton(2) - sqrt(2)) <= ulp(1.414)
-    """
-    return last(within(tol, iterate1(f, x0)))
-
 def interleave(*iterables):
     """Interleave items from several iterables. Generator.
 
@@ -1003,7 +909,8 @@ def powerset(iterable):
 def allsame(iterable):
     """Return whether all elements of an iterable are the same.
 
-    The test uses `!=` to compare.
+    The test uses `!=` to compare, and short-circuits at the
+    first item that is different.
 
     If `iterable` is empty, the return value is `True` (like for `all`).